+++ src/lib/libast/comp/wordexp.c
@@ -150,8 +150,8 @@ int	wordexp(const char *string, wordexp_t *wdarg, int flags)
 	if(flags&WRDE_DOOFFS)
 		c += wdarg->we_offs;
 	if(flags&WRDE_APPEND)
-		av = (char**)realloc(&wdarg->we_wordv[-1], (wdarg->we_wordc+c)*sizeof(char*));
-	else if(av = (char**)malloc(c*sizeof(char*)))
+		av = (char**)realloc(&wdarg->we_wordv[-1], (wdarg->we_wordc+(size_t)c)*sizeof(char*));
+	else if(av = (char**)malloc((size_t)c*sizeof(char*)))
 	{
 		if(flags&WRDE_DOOFFS)
 			memset(av,0,(wdarg->we_offs+1)*sizeof(char*));
@@ -161,7 +161,7 @@ int	wordexp(const char *string, wordexp_t *wdarg, int flags)
 	if(!av)
 		return WRDE_NOSPACE;
 	c = (int)stktell(stkstd);
-	if(!(cp = (char*)malloc(sizeof(char*)+c)))
+	if(!(cp = (char*)malloc(sizeof(char*)+(size_t)c)))
 	{
 		c=WRDE_NOSPACE;
 		goto err;
@@ -172,10 +172,10 @@ int	wordexp(const char *string, wordexp_t *wdarg, int flags)
 	wdarg->we_wordv = av;
 	if(flags&WRDE_APPEND)
 		av += wdarg->we_wordc;
-	wdarg->we_wordc += ac;
+	wdarg->we_wordc += (size_t)ac;
 	if(flags&WRDE_DOOFFS)
 		av += wdarg->we_offs;
-	memcpy(cp,stkptr(stkstd,offset),c);
+	memcpy(cp,stkptr(stkstd,offset),(size_t)c);
 	while(ac-- > 0)
 	{
 		*av++ = cp;
@@ -186,7 +186,7 @@ int	wordexp(const char *string, wordexp_t *wdarg, int flags)
 	c=0;
 err:
 	if(offset)
-		stkset(stkstd,savebase,offset);
+		stkset(stkstd,savebase,(size_t)offset);
 	else
 		stkseek(stkstd,0);
 	return c;
