+++ src/lib/libast/sfio/sfgetr.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2024 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 #include	"sfhdr.h"
@@ -80,7 +81,7 @@ char* sfgetr(Sfio_t*	f,	/* stream to read from	*/
 			}
 		}
 
-		if(!(s = (uchar*)memchr((char*)s,rc,n)))
+		if(!(s = (uchar*)memchr((char*)s,rc,(size_t)n)))
 			s = ends;
 	do_copy:
 		if(s < ends) /* found separator */
@@ -123,7 +124,7 @@ char* sfgetr(Sfio_t*	f,	/* stream to read from	*/
 		un += n;
 		ends = f->next;
 		f->next += n;
-		MEMCPY(s,ends,n);
+		MEMCPY(s,ends,(size_t)n);
 	}
 
 done:
