+++ src/lib/libast/tm/tmzone.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2012 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2023 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -103,7 +103,7 @@ tmzone(const char* name, char** end, const char* type, int* dst)
 		if (name[1] == 'E')
 			fixed.west = -fixed.west;
 		fixed.dst = name[0] == 'Z' ? 0 : d ? -d : TM_DST;
-		memcpy(fixed.standard = fixed.daylight = off, name, d);
+		memcpy(fixed.standard = fixed.daylight = off, name, (size_t)d);
 		off[d] = 0;
 		if (end)
 			*end = e;
